# feat: create an overload for `Any( SyntaxTriviaList` to prevent allocations